config and bean declarations fixed
